[factory-0.9] Switchboard Mac cmd line tool, Build from Cmd Line (#1145)
/apps/demos/ble-pairing/mac-client/build_client.sh

builds into:
/apps/demos/ble-pairing/mac-client/_build/mac-client

+++ apps/demos/ble-pairing/mac-client/mac-client/bleMessageProtocol.cpp
@@ -32,7 +32,7 @@ void Anki::Switchboard::BleMessageProtocol::ReceiveRawBuffer(uint8_t* buffer, si
   uint8_t mulipartState = BleMessageProtocol::GetMultipartBits(headerByte);
   
   if(sizeByte != size - 1) {
-    printf("Size failure: %d %d\n", sizeByte, size);
+    printf("Size failure: %d %d\n", sizeByte, (int)size);
     return;
   }
